# [misc] Use FASTVIDEO_STAGE_LOGGING for perf timing of stage