setting envvar `FASTVIDEO_STAGE_LOGGING=1` will result in the below logging:

```
NFO 07-25 01:23:24 [video_generator.py:224]                   flow_shift: 8
INFO 07-25 01:23:24 [video_generator.py:224]      embedded_guidance_scale: 6.0
INFO 07-25 01:23:24 [video_generator.py:224]                   save_video: True
INFO 07-25 01:23:24 [video_generator.py:224]                   output_path: video_samples_dmd2
INFO 07-25 01:23:24 [video_generator.py:224]         
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [composed_pipeline_base.py:102] Creating pipeline stages...
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [cuda.py:116] Trying FASTVIDEO_ATTENTION_BACKEND=VIDEO_SPARSE_ATTN
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [cuda.py:153] Using Video Sparse Attention backend.
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [composed_pipeline_base.py:311] Running pipeline stages: dict_keys(['input_validation_stage', 'prompt_encoding_stage', 'conditioning_stage', 'timestep_preparation_stage', 'latent_preparation_stage', 'denoising_stage', 'decoding_stage'])
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [InputValidationStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:156] [InputValidationStage] Execution completed in 0.03416999970795587 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [TextEncodingStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:156] [TextEncodingStage] Execution completed in 466.4717690029647 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [ConditioningStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:156] [ConditioningStage] Execution completed in 0.0208609999390319 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [TimestepPreparationStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:156] [TimestepPreparationStage] Execution completed in 0.4417329982970841 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [LatentPreparationStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:156] [LatentPreparationStage] Execution completed in 8.668422997288872 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:24 [base.py:150] [DmdDenoisingStage] Starting execution
  0%|                                                                                | 0/3 [00:00<?, ?it/s](FVWorkerProc-0 pid=282801) INFO 07-25 01:23:26 [config.py:58] PyTorch version 2.7.1+cu128 available.
100%|████████████████████████████████████████████████████████████████████████| 3/3 [00:03<00:00,  1.05s/it]
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:27 [base.py:156] [DmdDenoisingStage] Execution completed in 3167.356700003438 ms
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:27 [base.py:150] [DecodingStage] Starting execution
(FVWorkerProc-0 pid=282801) INFO 07-25 01:23:29 [base.py:156] [DecodingStage] Execution completed in 2210.2446550052264 ms
```